Revert "chore: FIT-671: Add state transitions to the current application flow…"
This reverts commit fc22d6f2fa978e0ce3aa185926ab63ccc3d8bc24.
