+++ label_studio/fsm/models.py
@@ -1,582 +1,349 @@
 """
-FSM Base Model for Label Studio.
-
-This module contains only FsmHistoryStateModel - the base class that models
-inherit from to get FSM integration. State model definitions are in state_models.py
-to avoid registration issues in LSE.
+Core FSM models for Label Studio.
 """
 
-import logging
-from typing import Any, Dict
+from datetime import datetime
+from typing import Any, Dict, Optional
 
+from django.conf import settings
 from django.db import models
-
-logger = logging.getLogger(__name__)
-
-
-# =============================================================================
-# FsmHistoryStateModel - Base Model for FSM Integration
-# =============================================================================
+from django.db.models import QuerySet, UUIDField
+from fsm.registry import register_state_model
+from fsm.state_choices import AnnotationStateChoices, ProjectStateChoices, TaskStateChoices
+from fsm.utils import UUID7Field, generate_uuid7, timestamp_from_uuid7
 
 
-class FsmHistoryStateModel(models.Model):
+class BaseState(models.Model):
     """
-    FSM History State Model - Base class for models that participate in FSM state tracking.
+    Abstract base class for all state models using UUID7 for optimal time-series performance.
 
-    This class provides explicit FSM integration through model lifecycle hooks,
-    replacing the implicit signal-based approach with predictable, testable behavior.
+    This is the core of the FSM system, providing:
+    - UUID7 primary key with natural time ordering
+    - Standard state transition metadata
+    - Audit trail information
+    - Context data storage
+    - Performance-optimized helper methods
 
-    Key features:
-    - Intercepts save operations to trigger FSM transitions
-    - Tracks field changes for transition logic
-    - Maintains CurrentContext for user/org tracking
-    - Provides explicit transition determination
-    - Fails gracefully - FSM errors don't break saves
-
-    Usage:
-        class Task(FsmHistoryStateModel):
-            # ... model fields ...
-
-            def _determine_fsm_transition(self) -> Optional[str]:
-                if self._state.adding:  # Creating new instance
-                    return 'task_created'
-
-                changed = self._get_changed_fields()
-                if 'is_labeled' in changed and changed['is_labeled'][1]:
-                    return 'task_labeled'
-
-                return None
-
-            def _get_fsm_transition_data(self) -> Dict[str, Any]:
-                return {
-                    'project_id': self.project_id,
-                    'overlap': self.overlap
-                }
+    Benefits of this architecture:
+    - INSERT-only operations for maximum concurrency
+    - Natural time ordering eliminates need for created_at indexes
+    - Global uniqueness enables distributed system support
+    - Time-based partitioning for large amounts of state records with consistent performance
+    - Complete audit trail by design
     """
+
+    # UUID7 Primary Key - provides natural time ordering and global uniqueness
+    id = UUIDField(
+        primary_key=True,
+        default=generate_uuid7,
+        editable=False,
+        help_text='UUID7 provides natural time ordering and global uniqueness',
+    )
+
+    # Optional organization field - can be overridden or left null
+    # Applications can add their own organization/tenant fields as needed
+    organization_id = models.PositiveIntegerField(
+        null=True,
+        blank=True,
+        db_index=True,
+        help_text='Organization ID that owns this state record (for multi-tenant applications)',
+    )
+
+    # Core State Fields
+    state = models.CharField(max_length=50, db_index=True, help_text='Current state of the entity')
+    previous_state = models.CharField(
+        max_length=50, null=True, blank=True, help_text='Previous state before this transition'
+    )
+
+    # Transition Metadata
+    transition_name = models.CharField(
+        max_length=100,
+        null=True,
+        blank=True,
+        help_text='Name of the transition method that triggered this state change',
+    )
+    triggered_by = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.SET_NULL,
+        null=True,
+        help_text='User who triggered this state transition',
+    )
+
+    # Context & Audit
+    context_data = models.JSONField(
+        default=dict, help_text='Additional context data for this transition (e.g., validation results, external IDs)'
+    )
+    reason = models.TextField(blank=True, help_text='Human-readable reason for this state transition')
+
+    # Timestamp (redundant with UUID7 but useful for human readability)
+    created_at = models.DateTimeField(
+        auto_now_add=True,
+        db_index=False,  # UUID7 provides natural ordering, no index needed
+        help_text='Human-readable timestamp for debugging (UUID7 id contains precise timestamp)',
+    )
 
     class Meta:
         abstract = True
+        # UUID7 provides natural ordering, reducing index requirements
+        ordering = ['-id']  # Most recent first
+        get_latest_by = 'id'
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Track original field values for change detection
-        # Initialize as empty dict for safe access
-        self._original_values = {}
+    def __str__(self):
+        entity_id = getattr(self, f'{self._get_entity_name()}_id', 'unknown')
+        return f'{self._get_entity_name().title()} {entity_id}: {self.previous_state} → {self.state}'
+
+    @property
+    def entity(self):
+        """Get the related entity object"""
+        entity_name = self._get_entity_name()
+        return getattr(self, entity_name)
+
+    @property
+    def timestamp_from_uuid(self) -> datetime:
+        """Extract timestamp from UUID7 ID"""
+        return timestamp_from_uuid7(self.id)
+
+    @property
+    def is_terminal_state(self) -> bool:
+        """
+        Check if this is a terminal state (no outgoing transitions).
+
+        Override in subclasses with specific terminal states.
+        """
+        return False
+
+    def _get_entity_name(self) -> str:
+        """Extract entity name from model name (e.g., TaskState → task)"""
+        model_name = self.__class__.__name__
+        if model_name.endswith('State'):
+            return model_name[:-5].lower()
+        return 'entity'
 
     @classmethod
-    def from_db(cls, db, field_names, values):
+    def get_current_state(cls, entity) -> Optional['BaseState']:
         """
-        Override from_db to capture original values when loading from database.
+        Get current state using UUID7 natural ordering.
 
-        Django calls this method instead of __init__ when loading models from the database.
-        We need to capture the original field values here for change detection.
+        Uses UUID7's natural time ordering to efficiently find the latest state
+        without requiring created_at indexes or complex queries.
         """
-        instance = super().from_db(db, field_names, values)
-        # Initialize as empty dict for safe access
-        instance._original_values = {}
-        # Capture original values immediately after loading from DB
-        # This ensures we have the baseline for change detection on the first save
-        instance._capture_original_values()
-        return instance
+        entity_field = f'{cls._get_entity_field_name()}'
+        return cls.objects.filter(**{entity_field: entity}).order_by('-id').first()
 
-    def _capture_original_values(self):
+    @classmethod
+    def get_current_state_value(cls, entity) -> Optional[str]:
         """
-        Capture current field values for change detection.
+        Get current state value as string using UUID7 natural ordering.
 
-        This allows us to detect which fields changed during save operations,
-        which is crucial for determining appropriate FSM transitions.
-
-        For ForeignKey fields, we store the PK instead of the object to avoid
-        circular references and recursion issues.
-
-        Deferred fields (not yet loaded from DB) are skipped to prevent infinite
-        recursion when accessing them would trigger refresh_from_db().
-
-        This is called after each save to refresh the baseline for the next save.
+        Uses UUID7's natural time ordering to efficiently find the latest state
+        without requiring created_at indexes or complex queries.
         """
-        self._original_values = {}
+        entity_field = f'{cls._get_entity_field_name()}'
+        current_state = cls.objects.filter(**{entity_field: entity}).order_by('-id').first()
+        return current_state.state if current_state else None
 
-        # Get deferred fields to avoid triggering recursive database loads
-        # Deferred fields haven't been loaded yet, so they can't have changed
-        deferred_fields = self.get_deferred_fields()
+    @classmethod
+    def get_state_history(cls, entity, limit: int = 100) -> QuerySet['BaseState']:
+        """Get complete state history for an entity"""
+        entity_field = f'{cls._get_entity_field_name()}'
+        return cls.objects.filter(**{entity_field: entity}).order_by('-id')[:limit]
 
-        for field in self._meta.fields:
-            # Skip deferred fields to prevent recursion via refresh_from_db()
-            if field.attname in deferred_fields:
-                continue
-
-            value = getattr(self, field.name, None)
-            # For ForeignKey fields, store PK to avoid circular references
-            if field.is_relation and field.many_to_one and value is not None:
-                self._original_values[field.name] = value.pk if hasattr(value, 'pk') else value
-            else:
-                self._original_values[field.name] = value
-
-    def __reduce_ex__(self, protocol):
+    @classmethod
+    def get_states_in_range(cls, entity, start_time: datetime, end_time: datetime) -> QuerySet['BaseState']:
         """
-        Override serialization to exclude internal FSM tracking fields.
+        Efficient time-range queries using UUID7.
 
-        Django's serialization uses pickle which calls __reduce_ex__.
-        We exclude _original_values since it's only needed for runtime
-        change detection, not for serialization/restoration.
+        Uses UUID7's embedded timestamp for direct time-based filtering
+        without requiring timestamp indexes.
         """
-        # Get the default reduction
-        reduction = super().__reduce_ex__(protocol)
+        entity_field = f'{cls._get_entity_field_name()}'
+        queryset = cls.objects.filter(**{entity_field: entity})
+        return UUID7Field.filter_by_time_range(queryset, start_time, end_time).order_by('id')
 
-        # reduction is a tuple: (callable, args, state, ...)
-        # state is the instance __dict__
-        if len(reduction) >= 3 and isinstance(reduction[2], dict):
-            state = reduction[2].copy()
-            # Remove internal FSM fields from serialization
-            state.pop('_original_values', None)
-            # Return new reduction with cleaned state
-            return (reduction[0], reduction[1], state) + reduction[3:]
+    @classmethod
+    def get_states_since(cls, entity, since: datetime):
+        """Get all states since a specific timestamp"""
+        entity_field = f'{cls._get_entity_field_name()}'
+        queryset = cls.objects.filter(**{entity_field: entity})
+        return UUID7Field.filter_since_time(queryset, since).order_by('id')
 
-        return reduction
-
-    def _get_changed_fields(self) -> Dict[str, tuple]:
+    @classmethod
+    def get_denormalized_fields(cls, entity) -> Dict[str, Any]:
         """
-        Get fields that changed since the last load/save.
+        Get denormalized fields to include in the state record.
 
-        Returns:
-            Dict mapping field names to (old_value, new_value) tuples
-            Note: For ForeignKey fields, old_value will be the PK, new_value will be the object
-
-        Example:
-            changed = self._get_changed_fields()
-            if 'is_labeled' in changed:
-                old_val, new_val = changed['is_labeled']
-                if not old_val and new_val:
-                    # Task became labeled
-                    pass
-        """
-        # If no original values captured yet, nothing has changed
-        # Use hasattr check to handle cases where _original_values doesn't exist
-        if not hasattr(self, '_original_values') or not self._original_values:
-            return {}
-
-        changed = {}
-        for field in self._meta.fields:
-            # Only check fields that were captured in _original_values
-            # Fields that were deferred during capture won't be in _original_values
-            # and should be considered unchanged
-            if field.name not in self._original_values:
-                continue
-
-            old_value = self._original_values[field.name]
-            new_value = getattr(self, field.name, None)
-
-            # For ForeignKey fields, old_value is stored as PK, so compare PK to PK
-            if field.is_relation and field.many_to_one:
-                new_pk = new_value.pk if new_value and hasattr(new_value, 'pk') else new_value
-                if old_value != new_pk:
-                    changed[field.name] = (old_value, new_value)
-            elif old_value != new_value:
-                changed[field.name] = (old_value, new_value)
-        return changed
-
-    def _determine_fsm_transitions(self, is_creating: bool = None, changed_fields: dict = None) -> list:
-        """
-        Determine which FSM transitions should be triggered based on model state.
-
-        This method automatically discovers registered transitions for this entity
-        and checks which ones should execute based on their trigger metadata.
+        Override this method in subclasses to provide denormalized data
+        that should be stored with each state transition for performance
+        optimization and auditing purposes.
 
         Args:
-            is_creating: Whether this is a creation. If None, checks self._state.adding
-            changed_fields: Dict of changed fields. If None, computes them.
-
-        Override this method ONLY if you need custom transition logic beyond
-        what the declarative triggers provide.
+            entity: The entity instance being transitioned
 
         Returns:
-            List of transition names to execute (in order)
-
-        Note:
-            In most cases, you don't need to override this. Just register transitions
-            with appropriate trigger metadata using @register_state_transition decorator.
-
-        Example of custom override (if needed):
-            def _determine_fsm_transitions(self, is_creating=None, changed_fields=None) -> list:
-                # Get default transitions
-                transitions = super()._determine_fsm_transitions(is_creating, changed_fields)
-
-                # Add custom logic
-                if self.some_complex_condition():
-                    transitions.append('custom_transition')
-
-                return transitions
-        """
-        from fsm.registry import transition_registry
-
-        entity_name = self._meta.model_name
-
-        # Use provided is_creating, or fall back to checking _state.adding
-        if is_creating is None:
-            is_creating = self._state.adding
-
-        # Use provided changed_fields, or compute them
-        if changed_fields is None:
-            changed_fields = {} if is_creating else self._get_changed_fields()
-
-        # Debug logging for transition determination
-        if entity_name == 'project' and not is_creating:
-            logger.debug(
-                f'FSM: Determining transitions for {entity_name}',
-                extra={
-                    'entity_id': self.pk,
-                    'is_creating': is_creating,
-                    'changed_fields': list(changed_fields.keys()),
-                    'changed_fields_detail': changed_fields,
-                },
-            )
-
-        # Get all registered transitions for this entity
-        registered_transitions = transition_registry.get_transitions_for_entity(entity_name)
-        if not registered_transitions:
-            return []
-
-        transitions_to_execute = []
-
-        for transition_name, transition_class in registered_transitions.items():
-            # Check if this transition should execute based on trigger metadata
-            should_execute = False
-
-            # Check creation trigger
-            if is_creating and getattr(transition_class, '_triggers_on_create', False):
-                should_execute = True
-
-            # Check update triggers
-            elif not is_creating and getattr(transition_class, '_triggers_on_update', True):
-                trigger_fields = getattr(transition_class, '_trigger_fields', [])
-
-                # If no specific fields, check if transition has custom logic
-                if not trigger_fields:
-                    # Let the transition's should_execute method decide
-                    # We'll add it and let it validate later
-                    should_execute = True
-                else:
-                    # Check if any trigger fields changed
-                    for field in trigger_fields:
-                        if field in changed_fields:
-                            should_execute = True
-                            break
-
-            # If trigger metadata says we should execute, also check the transition's should_execute() method
-            if should_execute:
-                try:
-                    # Instantiate the transition to check should_execute() if it exists
-                    # We need a minimal context to check should_execute
-                    from fsm.transitions import TransitionContext
-
-                    # Create a temporary transition instance with full context
-                    # Convert changed_fields to the format expected by ModelChangeTransition
-                    formatted_changed_fields = {k: {'old': v[0], 'new': v[1]} for k, v in changed_fields.items()}
-
-                    # Create transition with all relevant data for should_execute() check
-                    temp_transition = transition_class(
-                        is_creating=is_creating, changed_fields=formatted_changed_fields
-                    )
-
-                    # Check if should_execute is overridden (not using the base implementation)
-                    # The base implementation always returns True, so we only check if it's been customized
-                    from fsm.transitions import BaseTransition
-
-                    should_execute_method = getattr(type(temp_transition), 'should_execute', None)
-                    base_should_execute = getattr(BaseTransition, 'should_execute', None)
-
-                    # Only call should_execute if it's been overridden in the subclass
-                    if should_execute_method and should_execute_method != base_should_execute:
-                        # Build a minimal context for should_execute check
-                        # NOTE: We skip getting current state here to avoid recursion issues
-                        # The actual state will be retrieved during transition execution
-                        context = TransitionContext(
-                            entity=self,
-                            current_user=None,  # Will be set properly during execution
-                            current_state_object=None,  # Skip to avoid recursion
-                            current_state=None,  # Skip to avoid recursion
-                            target_state=temp_transition.target_state,
-                            organization_id=getattr(self, 'organization_id', None),
-                        )
-
-                        # Call should_execute to do final filtering
-                        if not temp_transition.should_execute(context):
-                            should_execute = False
-                            logger.debug(
-                                f'FSM: Transition {transition_name} filtered out by should_execute()',
-                                extra={
-                                    'entity_type': entity_name,
-                                    'entity_id': self.pk,
-                                    'transition_name': transition_name,
-                                },
-                            )
-                except Exception as e:
-                    # If should_execute check fails, log but still add the transition
-                    # Let it fail during actual execution with proper error handling
-                    logger.debug(
-                        f'FSM: Error checking should_execute for {transition_name}: {e}',
-                        extra={
-                            'entity_type': entity_name,
-                            'entity_id': self.pk,
-                            'transition_name': transition_name,
-                            'error': str(e),
-                        },
-                    )
-
-            if should_execute:
-                transitions_to_execute.append(transition_name)
-
-        return transitions_to_execute
-
-    def _get_fsm_transition_data(self) -> Dict[str, Any]:
-        """
-        Get data to pass to the FSM transition.
-
-        Override in subclasses to provide transition-specific data that should
-        be stored in the state record's context_data field.
-
-        Returns:
-            Dictionary of data to pass to transition
+            Dictionary of field names to values that should be stored
+            in the state record
 
         Example:
-            def _get_fsm_transition_data(self) -> Dict[str, Any]:
+            @classmethod
+            def get_denormalized_fields(cls, entity):
                 return {
-                    'project_id': self.project_id,
-                    'completed_by_id': self.completed_by_id,
-                    'annotation_count': self.annotations.count()
+                    'project_id': entity.project_id,
+                    'organization_id': entity.project.organization_id,
+                    'task_type': entity.task_type,
+                    'priority': entity.priority
                 }
         """
         return {}
 
-    def _should_execute_fsm(self) -> bool:
-        """
-        Check if FSM processing should be executed.
+    @classmethod
+    def _get_entity_field_name(cls) -> str:
+        """Get the foreign key field name for the entity"""
+        model_name = cls.__name__
+        if model_name.endswith('State'):
+            return model_name[:-5].lower()
+        return 'entity'
 
-        Returns False if:
-        - Feature flag is disabled
-        - User context is unavailable (tests must set CurrentContext explicitly)
-        - Explicitly skipped via instance attribute
 
-        Returns:
-            True if FSM should execute, False otherwise
+# Core state models for basic Label Studio entities
 
-        Note:
-            CurrentContext is available in web requests and background jobs.
-            In tests, it must be set explicitly for the user/organization.
-        """
-        # Check for instance-level skip flag
-        if getattr(self, '_skip_fsm', False):
-            return False
 
-        # Use the centralized FSM enabled check from utils
-        # This handles feature flag and thread-local overrides
-        try:
-            from core.current_request import CurrentContext
-            from fsm.utils import is_fsm_enabled
+@register_state_model('task')
+class TaskState(BaseState):
+    """
+    Core task state tracking for Label Studio.
+    Provides basic task state management with:
+    - Simple 3-state workflow (CREATED → IN_PROGRESS → COMPLETED)
+    - High-performance queries with UUID7 ordering
+    """
 
-            # Get user from CurrentContext - don't fall back to AnonymousUser
-            # If no user in context (e.g., tests without explicit setup), return False
-            try:
-                user = CurrentContext.get_user()
-                user_type = type(user).__name__ if user else None
-                user_authenticated = getattr(user, 'is_authenticated', None) if user else None
-                logger.info(
-                    f'FSM check for {self.__class__.__name__}(id={getattr(self, "pk", None)}): '
-                    f'user_type={user_type}, authenticated={user_authenticated}'
-                )
-                if user is None:
-                    logger.info(f'FSM check: User is None, skipping FSM for {self.__class__.__name__}')
-                    return False
-                # Check if user is authenticated (not AnonymousUser)
-                if not user.is_authenticated:
-                    logger.info(
-                        f'FSM check: User {user_type} not authenticated, skipping FSM for {self.__class__.__name__}'
-                    )
-                    return False
-            except Exception:
-                # CurrentContext not available or no user set
-                # This is expected in tests that don't set up context
-                logger.info(f'FSM check: Exception getting user, skipping FSM for {self.__class__.__name__}')
-                return False
+    # Entity Relationship
+    task = models.ForeignKey('tasks.Task', related_name='fsm_states', on_delete=models.CASCADE)
 
-            return is_fsm_enabled(user=user)
-        except Exception as e:
-            logger.debug(f'FSM check failed: {e}')
-            return False
+    # Override state field to add choices constraint
+    state = models.CharField(max_length=50, choices=TaskStateChoices.choices, db_index=True)
 
-    def save(self, *args, **kwargs):
-        """
-        Override save to trigger FSM transitions based on model changes.
+    project_id = models.PositiveIntegerField(
+        db_index=True, help_text='From task.project_id - denormalized for performance'
+    )
 
-        This method:
-        1. Captures the current state (creating vs updating)
-        2. Performs the actual database save
-        3. Determines if an FSM transition is needed
-        4. Executes the transition if needed
-        5. Gracefully handles FSM errors without breaking the save
+    class Meta:
+        app_label = 'fsm'
+        indexes = [
+            # Critical: Latest state lookup (current state determined by latest UUID7 id)
+            # Index with DESC order explicitly supports ORDER BY id DESC queries
+            models.Index(fields=['task_id', '-id'], name='task_current_state_idx'),
+            # Reporting and filtering
+            models.Index(fields=['project_id', 'state', '-id'], name='task_project_state_idx'),
+            models.Index(fields=['organization_id', 'state', '-id'], name='task_org_reporting_idx'),
+            # History queries
+            models.Index(fields=['task_id', 'id'], name='task_history_idx'),
+        ]
+        # No constraints needed - INSERT-only approach
+        ordering = ['-id']
 
-        Args:
-            *args: Positional arguments passed to super().save()
-            **kwargs: Keyword arguments passed to super().save()
-                     Special kwarg: skip_fsm=True to bypass FSM processing
+    @classmethod
+    def get_denormalized_fields(cls, entity):
+        """Get denormalized fields for TaskState creation"""
+        return {
+            'project_id': entity.project_id,
+        }
 
-        Returns:
-            Whatever super().save() returns
-        """
-        # Check for explicit FSM skip flag
-        skip_fsm = kwargs.pop('skip_fsm', False)
+    @property
+    def is_terminal_state(self) -> bool:
+        """Check if this is a terminal task state"""
+        return self.state == TaskStateChoices.COMPLETED
 
-        # Also check CurrentContext for skip_fsm flag (for context manager usage)
-        if not skip_fsm:
-            from core.current_request import CurrentContext
 
-            skip_fsm = CurrentContext.get('skip_fsm', False)
+@register_state_model('annotation')
+class AnnotationState(BaseState):
+    """
+    Core annotation state tracking for Label Studio.
+    Provides basic annotation state management with:
+    - Simple 3-state workflow (DRAFT → SUBMITTED → COMPLETED)
+    """
 
-        # Check if this is a creation vs update
-        is_creating = self._state.adding
+    # Entity Relationship
+    annotation = models.ForeignKey('tasks.Annotation', on_delete=models.CASCADE, related_name='fsm_states')
 
-        # Capture changed fields before save (only for updates)
-        # Note: _original_values should already be populated by from_db() or previous save()
-        changed_fields = {} if is_creating else self._get_changed_fields()
+    # Override state field to add choices constraint
+    state = models.CharField(max_length=50, choices=AnnotationStateChoices.choices, db_index=True)
 
-        # Perform the actual save
-        result = super().save(*args, **kwargs)
+    # Denormalized fields for performance (avoid JOINs in common queries)
+    task_id = models.PositiveIntegerField(
+        db_index=True, help_text='From annotation.task_id - denormalized for performance'
+    )
+    project_id = models.PositiveIntegerField(
+        db_index=True, help_text='From annotation.task.project_id - denormalized for performance'
+    )
+    completed_by_id = models.PositiveIntegerField(
+        null=True, db_index=True, help_text='From annotation.completed_by_id - denormalized for performance'
+    )
 
-        # After successful save, trigger FSM transitions if enabled and not skipped
-        should_execute = not skip_fsm and self._should_execute_fsm()
+    class Meta:
+        app_label = 'fsm'
+        indexes = [
+            # Critical: Latest state lookup
+            models.Index(fields=['annotation_id', '-id'], name='anno_current_state_idx'),
+            # Filtering and reporting
+            models.Index(fields=['task_id', 'state', '-id'], name='anno_task_state_idx'),
+            models.Index(fields=['completed_by_id', 'state', '-id'], name='anno_user_report_idx'),
+            models.Index(fields=['project_id', 'state', '-id'], name='anno_project_report_idx'),
+        ]
+        ordering = ['-id']
 
-        logger.debug(
-            f'FSM check for {self.__class__.__name__} {self.pk}: skip_fsm={skip_fsm}, should_execute={should_execute}',
-            extra={
-                'entity_type': self.__class__.__name__,
-                'entity_id': self.pk,
-                'skip_fsm': skip_fsm,
-                'should_execute': should_execute,
-            },
-        )
-        if not skip_fsm and should_execute:
-            try:
-                # Pass is_creating and changed_fields that were captured before save()
-                transitions = self._determine_fsm_transitions(is_creating=is_creating, changed_fields=changed_fields)
-                logger.debug(
-                    f'FSM transitions determined for {self.__class__.__name__} {self.pk}: {transitions}',
-                    extra={
-                        'entity_type': self.__class__.__name__,
-                        'entity_id': self.pk,
-                        'transitions': transitions,
-                        'is_creating': is_creating,
-                    },
-                )
-                for transition_name in transitions:
-                    try:
-                        self._execute_fsm_transition(
-                            transition_name=transition_name, is_creating=is_creating, changed_fields=changed_fields
-                        )
-                    except Exception as e:
-                        # Log error for this specific transition but continue with others
-                        logger.error(
-                            f'FSM transition {transition_name} failed for {self.__class__.__name__} {self.pk}',
-                            extra={
-                                'event': 'fsm.transition_failed_on_save',
-                                'entity_type': self.__class__.__name__,
-                                'entity_id': self.pk,
-                                'transition_name': transition_name,
-                                'error': str(e),
-                                'is_creating': is_creating,
-                            },
-                            exc_info=True,
-                        )
-            except Exception as e:
-                # Log error in determining transitions
-                logger.error(
-                    f'FSM transition discovery failed for {self.__class__.__name__} {self.pk}',
-                    extra={
-                        'event': 'fsm.transition_discovery_failed',
-                        'entity_type': self.__class__.__name__,
-                        'entity_id': self.pk,
-                        'error': str(e),
-                        'is_creating': is_creating,
-                    },
-                    exc_info=True,
-                )
+    @classmethod
+    def get_denormalized_fields(cls, entity):
+        """Get denormalized fields for AnnotationState creation"""
+        return {
+            'task_id': entity.task.id,
+            'project_id': entity.task.project_id,
+            'completed_by_id': entity.completed_by_id if entity.completed_by_id else None,
+        }
 
-        # Update original values after save for next time
-        self._capture_original_values()
+    @property
+    def is_terminal_state(self) -> bool:
+        """Check if this is a terminal annotation state"""
+        return self.state == AnnotationStateChoices.COMPLETED
 
-        return result
 
-    def _execute_fsm_transition(self, transition_name: str, is_creating: bool, changed_fields: Dict[str, tuple]):
-        """
-        Execute an FSM transition.
+@register_state_model('project')
+class ProjectState(BaseState):
+    """
+    Core project state tracking for Label Studio.
+    Provides basic project state management with:
+    - Simple 3-state workflow (CREATED → IN_PROGRESS → COMPLETED)
+    - Project lifecycle tracking
+    """
 
-        This method handles the actual transition execution, including:
-        - Getting current context (user, org_id)
-        - Preparing transition data
-        - Calling the state manager
+    # Entity Relationship
+    project = models.ForeignKey('projects.Project', on_delete=models.CASCADE, related_name='fsm_states')
 
-        Args:
-            transition_name: Name of the registered transition to execute
-            is_creating: Whether this is a new model creation
-            changed_fields: Dict of changed fields (field_name -> (old, new))
+    # Override state field to add choices constraint
+    state = models.CharField(max_length=50, choices=ProjectStateChoices.choices, db_index=True)
 
-        Note:
-            This is only called after _should_execute_fsm() returns True,
-            so CurrentContext should be available with a valid user.
-        """
-        from core.current_request import CurrentContext
-        from fsm.state_manager import get_state_manager
+    created_by_id = models.PositiveIntegerField(
+        null=True, db_index=True, help_text='From project.created_by_id - denormalized for performance'
+    )
 
-        StateManager = get_state_manager()
+    class Meta:
+        app_label = 'fsm'
+        indexes = [
+            # Critical: Latest state lookup
+            models.Index(fields=['project_id', '-id'], name='project_current_state_idx'),
+            # Filtering and reporting
+            models.Index(fields=['organization_id', 'state', '-id'], name='project_org_state_idx'),
+            models.Index(fields=['organization_id', '-id'], name='project_org_reporting_idx'),
+        ]
+        ordering = ['-id']
 
-        # Get context - should be available since _should_execute_fsm passed
-        user = CurrentContext.get_user()
-        org_id = CurrentContext.get_organization_id()
+    @classmethod
+    def get_denormalized_fields(cls, entity):
+        """Get denormalized fields for ProjectState creation"""
+        return {
+            'created_by_id': entity.created_by_id if entity.created_by_id else None,
+        }
 
-        # Get transition-specific data from the model
-        transition_data = self._get_fsm_transition_data()
-
-        # Add metadata about the change
-        transition_data.update(
-            {
-                'is_creating': is_creating,
-                'changed_fields': {k: {'old': v[0], 'new': v[1]} for k, v in changed_fields.items()},
-            }
-        )
-
-        logger.info(
-            f'Executing FSM transition for {self.__class__.__name__}',
-            extra={
-                'event': 'fsm.transition_executing',
-                'entity_type': self.__class__.__name__,
-                'entity_id': self.pk,
-                'transition_name': transition_name,
-                'is_creating': is_creating,
-                'user_id': user.id if user else None,
-                'organization_id': org_id,
-            },
-        )
-
-        # Execute the registered transition
-        try:
-            StateManager.execute_transition(
-                entity=self,
-                transition_name=transition_name,
-                transition_data=transition_data,
-                user=user,
-                organization_id=org_id,
-            )
-
-            logger.info(
-                f'FSM transition executed successfully for {self.__class__.__name__}',
-                extra={
-                    'event': 'fsm.transition_success',
-                    'entity_type': self.__class__.__name__,
-                    'entity_id': self.pk,
-                    'transition_name': transition_name,
-                    'user_id': user.id if user else None,
-                    'organization_id': org_id,
-                },
-            )
-        except Exception:
-            # Re-raise to be caught by save() method
-            raise
+    @property
+    def is_terminal_state(self) -> bool:
+        """Check if this is a terminal project state"""
+        return self.state == ProjectStateChoices.COMPLETED
