+++ label_studio/fsm/state_manager.py
@@ -9,13 +9,12 @@ import logging
 from datetime import datetime
 from typing import Any, Dict, List, Optional, Type
 
-from core.current_request import CurrentContext
-from core.feature_flags import flag_set
 from django.conf import settings
 from django.core.cache import cache
+from django.db import transaction
 from django.db.models import Model, QuerySet
+from fsm.models import BaseState
 from fsm.registry import get_state_model_for_entity
-from fsm.state_models import BaseState
 from fsm.transition_executor import execute_transition_with_state_manager
 
 logger = logging.getLogger(__name__)
@@ -55,13 +54,6 @@ class StateManager:
     CACHE_PREFIX = 'fsm:current'
 
     @classmethod
-    def _is_fsm_enabled(cls, user='auto') -> bool:
-        if user == 'auto':
-            user = CurrentContext.get_user()
-        """Check if FSM feature is enabled via feature flag."""
-        return flag_set('fflag_feat_fit_568_finite_state_management', user=user)
-
-    @classmethod
     def get_cache_key(cls, entity: Model) -> str:
         """Generate cache key for entity's current state"""
         return f'{cls.CACHE_PREFIX}:{entity._meta.label_lower}:{entity.pk}'
@@ -86,21 +78,17 @@ class StateManager:
                 # Task is finished
                 pass
         """
-        if not cls._is_fsm_enabled():
-            return None  # Feature disabled, return no state
-
         cache_key = cls.get_cache_key(entity)
 
         # Try cache first
         cached_state = cache.get(cache_key)
         if cached_state is not None:
             logger.info(
-                'FSM: Cache hit',
+                'FSM cache hit',
                 extra={
                     'event': 'fsm.cache_hit',
                     'entity_type': entity._meta.label_lower,
                     'entity_id': entity.pk,
-                    'organization_id': CurrentContext.get_organization_id(),
                     'state': cached_state,
                 },
             )
@@ -118,12 +106,11 @@ class StateManager:
             if current_state is not None:
                 cache.set(cache_key, current_state, cls.CACHE_TTL)
                 logger.info(
-                    'FSM: Cache miss',
+                    'FSM cache miss',
                     extra={
                         'event': 'fsm.cache_miss',
                         'entity_type': entity._meta.label_lower,
                         'entity_id': entity.pk,
-                        'organization_id': CurrentContext.get_organization_id(),
                     },
                 )
 
@@ -131,12 +118,11 @@ class StateManager:
 
         except Exception as e:
             logger.error(
-                'FSM: Error getting current state',
+                'Error getting current state',
                 extra={
                     'event': 'fsm.get_state_error',
                     'entity_type': entity._meta.label_lower,
                     'entity_id': entity.pk,
-                    'organization_id': CurrentContext.get_organization_id(),
                     'error': str(e),
                 },
                 exc_info=True,
@@ -172,10 +158,8 @@ class StateManager:
         new_state: str,
         transition_name: str = None,
         user=None,
-        organization_id=None,
         context: Dict[str, Any] = None,
         reason: str = '',
-        force_state_record: bool = False,
     ) -> bool:
         """
         Perform state transition with audit trail.
@@ -190,10 +174,8 @@ class StateManager:
             new_state: Target state
             transition_name: Name of transition method (for audit)
             user: User triggering the transition
-            organization_id: Organization ID
             context: Additional context data
             reason: Human-readable reason for transition
-            force_state_record: If True, creates state record even if state doesn't change (for audit trails)
 
         Returns:
             True if transition succeeded, False otherwise
@@ -208,78 +190,32 @@ class StateManager:
                 new_state='IN_PROGRESS',
                 transition_name='start_annotation',
                 user=request.user,
-                organization_id=request.user.active_organization_id,
                 context={'assignment_id': assignment.id},
                 reason='User started annotation work'
             )
         """
-        if not cls._is_fsm_enabled(user=user):
-            return True  # Feature disabled, silently succeed
-
         state_model = get_state_model_for_entity(entity)
         if not state_model:
             raise StateManagerError(f'No state model found for {entity._meta.model_name} when transitioning state')
 
         current_state = cls.get_current_state_value(entity)
 
-        # Prevent same-state transitions - only create state records for actual state changes
-        # This avoids creating redundant data when the effective state doesn't change
-        # However, allow forced state records for audit trails (e.g., annotation updates)
-        # IMPORTANT: Also check if a state record exists in DB - if not, we must create one
-        # even if inferred state matches target state (to persist the inferred state)
-        if current_state == new_state and not force_state_record:
-            # Verify a state record actually exists in DB (not just inferred)
-            state_record_exists = state_model.objects.filter(**{entity._meta.model_name: entity}).exists()
-            if state_record_exists:
-                return True  # Skip transition - record exists and state unchanged
-            # else: No record exists (state was inferred), continue to create record
-
-        # Optimistic concurrency control using cache-based locking
-        cache_key = cls.get_cache_key(entity)
-        lock_key = f'{cache_key}:lock'
-
-        if organization_id is None:
-            organization_id = CurrentContext.get_organization_id()
-
         try:
-            # Try to acquire an optimistic lock using cache add (atomic operation)
-            # add() only succeeds if the key doesn't exist
-            lock_acquired = cache.add(lock_key, 'locked', timeout=5)  # 5 second timeout
-
-            if not lock_acquired:
-                # Another process is currently transitioning this entity
-                logger.info(
-                    'FSM: Concurrent transition detected, skipping',
-                    extra={
-                        'event': 'fsm.concurrent_transition_skipped',
-                        'entity_type': entity._meta.label_lower,
-                        'entity_id': entity.pk,
-                        'target_state': new_state,
-                        'organization_id': organization_id,
-                    },
-                )
-                return True
-
-            try:
+            with transaction.atomic():
                 # INSERT-only approach - no UPDATE operations needed
                 # Get denormalized fields from the state model class
                 denormalized_fields = state_model.get_denormalized_fields(entity)
 
                 # Get organization from entity or denormalized fields, or user's active organization
-                if organization_id is None:
-                    organization_id = getattr(
-                        entity, 'organization_id', getattr(denormalized_fields, 'organization_id', None)
-                    )
-                    if organization_id is not None:
-                        CurrentContext.set_organization_id(organization_id)
+                organization_id = getattr(
+                    entity, 'organization_id', getattr(denormalized_fields, 'organization_id', None)
+                )
 
                 if not organization_id and user and hasattr(user, 'active_organization') and user.active_organization:
                     organization_id = user.active_organization.id
-                    if organization_id is not None:
-                        CurrentContext.set_organization_id(organization_id)
 
                 logger.info(
-                    'FSM: State transition starting',
+                    'State transition starting',
                     extra={
                         'event': 'fsm.transition_state_start',
                         'entity_type': entity._meta.label_lower,
@@ -294,13 +230,8 @@ class StateManager:
                     },
                 )
 
-                # CRITICAL FIX: Use state model's correct field name instead of entity._meta.model_name
-                # This fixes the architectural entity field mapping issue where entity._meta.model_name
-                # doesn't always match the actual field name defined in FSM state models
-                entity_field_name = state_model._get_entity_field_name()
-
                 new_state_record = state_model.objects.create(
-                    **{entity_field_name: entity},
+                    **{entity._meta.model_name: entity},
                     state=new_state,
                     previous_state=current_state,
                     transition_name=transition_name,
@@ -311,26 +242,28 @@ class StateManager:
                     **denormalized_fields,
                 )
 
-                # Write-through cache: Update immediately
-                # This ensures the cache is updated atomically with the database
-                cache.set(cache_key, new_state, cls.CACHE_TTL)
+                # Update cache with new state after transaction commits
+                cache_key = cls.get_cache_key(entity)
 
-                logger.info(
-                    'FSM: Cache updated for transition state',
-                    extra={
-                        'event': 'fsm.transition_state_cache_updated',
-                        'entity_type': entity._meta.label_lower,
-                        'entity_id': entity.pk,
-                        'state': new_state,
-                        **{
-                            'user_id': user.id if user else None,
-                            'organization_id': organization_id if organization_id else None,
+                def update_cache(key, state, user_id, org_id):
+                    cache.set(key, state, cls.CACHE_TTL)
+                    logger.info(
+                        'Cache updated for transition state',
+                        extra={
+                            'event': 'fsm.transition_state_cache_updated',
+                            'entity_type': entity._meta.label_lower,
+                            'entity_id': entity.pk,
+                            'state': state,
+                            **{'user_id': user_id if user_id else None, 'organization_id': org_id if org_id else None},
                         },
-                    },
+                    )
+
+                transaction.on_commit(
+                    lambda: update_cache(cache_key, new_state, user.id if user else None, organization_id)
                 )
 
                 logger.info(
-                    'FSM: State transition successful',
+                    'State transition successful',
                     extra={
                         'event': 'fsm.transition_state_success',
                         'entity_type': entity._meta.label_lower,
@@ -345,20 +278,12 @@ class StateManager:
                 )
                 return True
 
-            finally:
-                # Always release the lock, regardless of success or failure
-                cache.delete(lock_key)
-
         except Exception as e:
-            # On failure, clean up lock and invalidate potentially stale cache
-            cache.delete(lock_key)
+            # On failure, invalidate potentially stale cache
+            cache_key = cls.get_cache_key(entity)
             cache.delete(cache_key)
-
-            # Get organization_id for error logging if it wasn't set earlier
-            organization_id = CurrentContext.get_organization_id()
-
             logger.error(
-                'FSM: State transition failed',
+                'State transition failed',
                 extra={
                     'event': 'fsm.transition_state_failed',
                     'entity_type': entity._meta.label_lower,
@@ -423,9 +348,9 @@ class StateManager:
         """Invalidate cached state for an entity"""
         cache_key = cls.get_cache_key(entity)
         cache.delete(cache_key)
-        organization_id = CurrentContext.get_organization_id()
+        organization_id = getattr(entity, 'organization_id', None)
         logger.info(
-            'FSM: Cache invalidated',
+            'Cache invalidated',
             extra={
                 'event': 'fsm.cache_invalidated',
                 'entity_type': entity._meta.label_lower,
@@ -443,8 +368,11 @@ class StateManager:
         bulk queries and advanced caching strategies.
         """
         cache_updates = {}
-        organization_id = CurrentContext.get_organization_id()
+        organization_id = None
         for entity in entities:
+            if organization_id is None:
+                if hasattr(entity, 'organization_id'):
+                    organization_id = entity.organization_id
             current_state = cls.get_current_state_value(entity)
             if current_state:
                 cache_key = cls.get_cache_key(entity)
@@ -453,7 +381,7 @@ class StateManager:
         if cache_updates:
             cache.set_many(cache_updates, cls.CACHE_TTL)
             logger.info(
-                'FSM: Cache warmed',
+                'Cache warmed',
                 extra={
                     'event': 'fsm.cache_warmed',
                     'entity_count': len(cache_updates),
