+++ label_studio/pytest.ini
@@ -8,5 +8,3 @@ env =
     D:SENTRY_DSN=
     D:USE_ENFORCE_CSRF_CHECKS=0
     D:TEST_ENVIRONMENT=1
-    COLLECT_ANALYTICS=0
-    fflag_feat_fit_568_finite_state_management=1
